+++ tests/UnitTestApp/UnitTestApp/SharedMockFiles/Categories/ANBannerAdView+ANTest.m
@@ -14,11 +14,16 @@
  */
 
 #import "ANBannerAdView+ANTest.h"
+#import "NSObject+Swizzling.h"
+#import <objc/runtime.h>
 
 #pragma clang diagnostic push
 #pragma clang diagnostic ignored "-Wprotocol"
 #pragma clang diagnostic ignored "-Wincomplete-implementation"
 @implementation ANBannerAdView (ANTest)
+
+static BOOL kDoNotResetAdUnitUUIDEnabled = NO;
+
 #pragma clang diagnostic pop
 
 @dynamic placementId;
@@ -30,5 +35,25 @@
 @dynamic customKeywords;
 @dynamic contentView;
 @dynamic transitionInProgress;
+
++ (void)setDoNotResetAdUnitUUID:(BOOL)simulationEnabled {
+    kDoNotResetAdUnitUUIDEnabled = simulationEnabled;
+}
+
++ (void)load {
+    NSBlockOperation *operation = [NSBlockOperation blockOperationWithBlock:^{
+        [[self class] exchangeInstanceSelector:@selector(internalUTRequestUUIDStringReset)
+                                  withSelector:@selector(swizzle_internalUTRequestUUIDStringReset)];
+    }];
+    [operation start];
+}
+
+- (void)swizzle_internalUTRequestUUIDStringReset {
+    if(kDoNotResetAdUnitUUIDEnabled){
+        NSLog(@"Do nothing");
+    }else{
+        [self swizzle_internalUTRequestUUIDStringReset];
+    }
+}
 
 @end