+++ RELEASE-NOTES.md
@@ -1,3 +1,16 @@
+## 7.0
+
+### New Feature
++ MS-3885 Added support for Multi Ad Request (MAR) (https://wiki.xandr.com/display/sdk/Single+Request+Mode+for+iOS
+
+### Mediation partner upgrade
++ Smart Server SDK v7.3.0
+
+### Improvements/Bug Fixes
++ MS-4122 - Fixed potential memory leak issue caused by improper deallocation of InstreamVideoAd object
++ MS-4094 - Code clean up to fix issue with OMID SDK video implementation
+
+
 ## 6.2
 
 ### New Features